Update all dependencies

+++ shader/src/lib.rs
@@ -32,7 +32,7 @@ use std::collections::HashMap;
 ///
 pub trait Shader {
     /// Get spirv bytecode.
-    fn spirv(&self) -> Result<std::borrow::Cow<'_, [u8]>, failure::Error>;
+    fn spirv(&self) -> Result<std::borrow::Cow<'_, [u32]>, failure::Error>;
 
     /// Get the entry point of the shader.
     fn entry(&self) -> &str;
@@ -61,14 +61,14 @@ pub trait Shader {
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct SpirvShader {
     #[cfg_attr(feature = "serde", serde(with = "serde_bytes"))]
-    spirv: Vec<u8>,
+    spirv: Vec<u32>,
     stage: ShaderStageFlags,
     entry: String,
 }
 
 impl SpirvShader {
     /// Create Spir-V shader from bytes.
-    pub fn new(spirv: Vec<u8>, stage: ShaderStageFlags, entrypoint: &str) -> Self {
+    pub fn new(spirv: Vec<u32>, stage: ShaderStageFlags, entrypoint: &str) -> Self {
         assert!(!spirv.is_empty());
         assert_eq!(spirv.len() % 4, 0);
         Self {
@@ -80,7 +80,7 @@ impl SpirvShader {
 }
 
 impl Shader for SpirvShader {
-    fn spirv(&self) -> Result<std::borrow::Cow<'_, [u8]>, failure::Error> {
+    fn spirv(&self) -> Result<std::borrow::Cow<'_, [u32]>, failure::Error> {
         Ok(std::borrow::Cow::Borrowed(&self.spirv))
     }
 
@@ -172,12 +172,12 @@ pub struct SpecConstantSet {
 #[derive(Clone, Debug, Default)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct ShaderSetBuilder {
-    vertex: Option<(Vec<u8>, String)>,
-    fragment: Option<(Vec<u8>, String)>,
-    geometry: Option<(Vec<u8>, String)>,
-    hull: Option<(Vec<u8>, String)>,
-    domain: Option<(Vec<u8>, String)>,
-    compute: Option<(Vec<u8>, String)>,
+    vertex: Option<(Vec<u32>, String)>,
+    fragment: Option<(Vec<u32>, String)>,
+    geometry: Option<(Vec<u32>, String)>,
+    hull: Option<(Vec<u32>, String)>,
+    domain: Option<(Vec<u32>, String)>,
+    compute: Option<(Vec<u32>, String)>,
 }
 
 impl ShaderSetBuilder {
@@ -200,7 +200,7 @@ impl ShaderSetBuilder {
         }
 
         let create_storage = move |stage,
-                                   shader: (Vec<u8>, String, Option<gfx_hal::pso::Specialization<'static>>),
+                                   shader: (Vec<u32>, String, Option<gfx_hal::pso::Specialization<'static>>),
                                    factory|
               -> Result<ShaderStorage<B>, failure::Error> {
             let mut storage = ShaderStorage {
@@ -347,7 +347,7 @@ impl ShaderSetBuilder {
 #[derive(Debug)]
 pub struct ShaderStorage<B: Backend> {
     stage: ShaderStageFlags,
-    spirv: Vec<u8>,
+    spirv: Vec<u32>,
     module: Option<B::ShaderModule>,
     entrypoint: String,
     specialization: Option<gfx_hal::pso::Specialization<'static>>,
